+++ BitmovinAnalyticsCollector/Classes/BitmovinPlayer/Classes/BitmovinAdAdapter.swift
@@ -12,16 +12,16 @@ public class BitmovinAdAdapter : NSObject, AdAdapter {
         self.bitmovinPlayer.add(listener: self)
     }
 
-    func releaseAdapter() {
+    public func releaseAdapter() {
         self.bitmovinPlayer.remove(listener: self)
     }
     
-    func getModuleInformation() -> AdModuleInformation {
+    public func getModuleInformation() -> AdModuleInformation {
         let playerVersion = BitmovinPlayerUtil.playerVersion()
         return AdModuleInformation(name: "BMPDefaultAdvertisingService", version: playerVersion)
     }
     
-    func isAutoPlayEnabled() -> Bool {
+    public func isAutoPlayEnabled() -> Bool {
         self.bitmovinPlayer.config.playbackConfiguration.isAutoplayEnabled
     }
 }