+++ BitmovinAnalyticsCollector/Classes/Collector/ads/AnalyticsAdBreak.swift
@@ -1,19 +1,19 @@
 import Foundation
 
 public class AnalyticsAdBreak {
-    var id: String
-    var ads: Array<AnalyticsAd>
-    var position: AdPosition?
-    var offset: String?
-    var scheduleTime: TimeInterval?
-    var replaceContentDuration: TimeInterval?
-    var preloadOffset: Int64?
-    var tagType: AnalyticsAdTagType?
-    var tagUrl: String?
-    var persistent: Bool?
-    var fallbackIndex: Int = 0
+    public var id: String
+    public var ads: Array<AnalyticsAd>
+    public var position: AdPosition?
+    public var offset: String?
+    public var scheduleTime: TimeInterval?
+    public var replaceContentDuration: TimeInterval?
+    public var preloadOffset: Int64?
+    public var tagType: AnalyticsAdTagType?
+    public var tagUrl: String?
+    public var persistent: Bool?
+    public var fallbackIndex: Int = 0
     
-    init(id: String, ads: Array<AnalyticsAd>) {
+    public init(id: String, ads: Array<AnalyticsAd>) {
         self.id = id
         self.ads = ads
     }