+++ BitmovinAnalyticsCollector/Classes/Collector/data/AdModuleInformation.swift
@@ -4,7 +4,7 @@ public class AdModuleInformation {
     var name: String
     var version: String?
     
-    init(name: String, version: String?) {
+    public init(name: String, version: String?) {
         self.name = name
         self.version = version
     }