+++ BitmovinAnalyticsCollector/Classes/Collector/data/ErrorData.swift
@@ -9,7 +9,7 @@ public class ErrorData {
     static let ANALYTICS_BUFFERING_TIMEOUT_REACHED = ErrorData(code: 10001, message: "ANALYTICS_BUFFERING_TIMEOUT_REACHED", data: nil)
     static let ANALYTICS_VIDEOSTART_TIMEOUT_REACHED = ErrorData(code: 10002, message: "ANALYTICS_VIDEOSTART_TIMEOUT_REACHED", data: nil)
     
-    init(code: Int?, message: String?, data: String?) {
+    public init(code: Int?, message: String?, data: String?) {
         self.code = code
         self.message = message
         self.data = data