+++ BitmovinAnalyticsCollector/Tests/CoreTests/CoreTest.swift
@@ -1,5 +1,7 @@
 import XCTest
+#if !SWIFT_PACKAGE
 @testable import BitmovinAnalyticsCollector
+#endif
 
 class CoreTests_iOS: XCTestCase {