+++ BitmovinAnalyticsCollector/Tests/CoreTests/data/EventDataFactoryTests.swift
@@ -1,5 +1,11 @@
 import XCTest
+#if !SWIFT_PACKAGE
 @testable import BitmovinAnalyticsCollector
+#endif
+
+#if SWIFT_PACKAGE
+@testable import BitmovinCollectorCore
+#endif
 
 class EventDataFactoryTests: XCTestCase {