+++ BitmovinAnalyticsCollector/Tests/CoreTests/mock/FakePlayerAdapter.swift
@@ -1,5 +1,11 @@
 import CoreMedia
+#if !SWIFT_PACKAGE
 @testable import BitmovinAnalyticsCollector
+#endif
+
+#if SWIFT_PACKAGE
+@testable import BitmovinCollectorCore
+#endif
 
 class FakePlayerAdapter: PlayerAdapter {
     func decorateEventData(eventData: EventData) {