+++ BitmovinAnalyticsCollector/Tests/CoreTests/mock/MockStateMachineDelegate.swift
@@ -1,5 +1,11 @@
 import CoreMedia
+#if !SWIFT_PACKAGE
 @testable import BitmovinAnalyticsCollector
+#endif
+
+#if SWIFT_PACKAGE
+@testable import BitmovinCollectorCore
+#endif
 
 class MockStateMachineDelegate: StateMachineDelegate {
     func stateMachineDidExitSetup(_ stateMachine: StateMachine) {